requirements.txt
django
dj-database-url
psycopg2-binary
django-allauth
django-environ
split_settings
djangorestframework
Pillow
docutils
django-extensions
